Refactor project

+++ app/models/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime
+from sqlalchemy import DateTime, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.sql import func
 
@@ -10,7 +10,7 @@ from app.database import Base
 class Users(Base):
     __tablename__ = 'users'
 
-    user_id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True, autoincrement=True)
     username: Mapped[str]
     email: Mapped[str]
     hashed_password: Mapped[str]
